- Moved Active Assets out of freeroamEvents.lua
